starrailcard/requirements.txt
aiofiles
aiohttp
cachetools
imageio
moviepy
more-itertools
numpy
Pillow
pydantic
beautifulsoup4
